Refactor and update environments card and list pages
