+++ src/components/Environment/EnvironmentCard.tsx
@@ -9,21 +9,56 @@ import {
   TextContent,
   TextVariants,
   Label,
+  DescriptionList,
+  DescriptionListDescription,
+  DescriptionListGroup,
+  DescriptionListTerm,
 } from '@patternfly/react-core';
-import { useApplications } from '../../hooks/useApplications';
+import { useEnvironmentActions } from '../../components/Environment/environment-actions';
+import { EnvironmentKindWithHealthStatus } from '../../hooks/useAllApplicationEnvironmentsWithHealthStatus';
 import ActionMenu from '../../shared/components/action-menu/ActionMenu';
+import { Timestamp } from '../../shared/components/timestamp/Timestamp';
 import { EnvironmentKind } from '../../types';
 import { getEnvironmentDeploymentStrategyLabel } from '../../utils/environment-utils';
-import { useWorkspaceInfo } from '../../utils/workspace-context-utils';
-import { useEnvironmentActions } from './environment-actions';
+import { getGitOpsDeploymentHealthStatusIcon } from '../../utils/gitops-utils';
+import { EnvironmentType, getEnvironmentType, getEnvironmentTypeLabel } from './environment-utils';
+
+const ApplicationEnvironmentStatus: React.FC<{
+  environment: EnvironmentKindWithHealthStatus;
+}> = ({ environment }) => {
+  if (!environment.healthStatus) return null;
+
+  return (
+    <>
+      <DescriptionListGroup>
+        <DescriptionListTerm>Application status</DescriptionListTerm>
+        <DescriptionListDescription>
+          <Text component={TextVariants.small} style={{ color: 'var(--pf-global--Color--200)' }}>
+            {getGitOpsDeploymentHealthStatusIcon(environment.healthStatus)}{' '}
+            {environment.healthStatus}
+          </Text>
+        </DescriptionListDescription>
+      </DescriptionListGroup>
+      <DescriptionListGroup>
+        <DescriptionListTerm>Last deploy</DescriptionListTerm>
+        <DescriptionListDescription>
+          <Text component={TextVariants.small} style={{ color: 'var(--pf-global--Color--200)' }}>
+            <Timestamp timestamp={environment.lastDeploy} simple />
+          </Text>
+        </DescriptionListDescription>
+      </DescriptionListGroup>
+    </>
+  );
+};
 
 type EnvironmentCardProps = {
   environment: EnvironmentKind;
+  readOnly?: boolean;
 };
-const EnvironmentCard: React.FC<EnvironmentCardProps> = ({ environment }) => {
+
+const EnvironmentCard: React.FC<EnvironmentCardProps> = ({ environment, readOnly }) => {
   const actions = useEnvironmentActions(environment);
-  const { namespace } = useWorkspaceInfo();
-  const [applications, applicationsLoaded] = useApplications(namespace);
+  const type = getEnvironmentType(environment);
 
   return (
     <Card isFlat>
@@ -35,24 +70,40 @@ const EnvironmentCard: React.FC<EnvironmentCardProps> = ({ environment }) => {
             </Text>
           </TextContent>
         </CardTitle>
-        <CardActions>
-          <ActionMenu actions={actions} />
-        </CardActions>
+        {!readOnly && actions?.length ? (
+          <CardActions>
+            <ActionMenu actions={actions} />
+          </CardActions>
+        ) : null}
       </CardHeader>
       <CardBody>
-        <TextContent>
-          <Text>
-            <b>Deployment strategy:</b>
-          </Text>
-        </TextContent>
-        <Label>{getEnvironmentDeploymentStrategyLabel(environment)}</Label>
-      </CardBody>
-      <CardBody>
-        <TextContent>
-          <Text>
-            <b>Applications: {applicationsLoaded ? applications.length : '-'}</b>
-          </Text>
-        </TextContent>
+        <DescriptionList>
+          <DescriptionListGroup>
+            <DescriptionListTerm>Type</DescriptionListTerm>
+            <DescriptionListDescription>
+              <Label
+                color={
+                  type === EnvironmentType.managed
+                    ? 'green'
+                    : type === EnvironmentType.static
+                    ? 'gold'
+                    : 'cyan'
+                }
+              >
+                {getEnvironmentTypeLabel(type)}
+              </Label>
+            </DescriptionListDescription>
+          </DescriptionListGroup>
+          <DescriptionListGroup>
+            <DescriptionListTerm>Deployment strategy</DescriptionListTerm>
+            <DescriptionListDescription>
+              <Label>{getEnvironmentDeploymentStrategyLabel(environment)}</Label>
+            </DescriptionListDescription>
+          </DescriptionListGroup>
+          <ApplicationEnvironmentStatus
+            environment={environment as EnvironmentKindWithHealthStatus}
+          />
+        </DescriptionList>
       </CardBody>
     </Card>
   );
