+++ src/components/Environment/EnvironmentList.scss
@@ -1,4 +1,4 @@
-.environment-list-view {
+.environment-list {
   &_card .pf-c-card {
     height: 100%;
   }
