--- src/components/WorkspaceSettings/WorkspaceSettings.scss
@@ -1,9 +1,0 @@
-.workspace-settings {
-  &__card-img {
-    --pf-u-min-width--MinWidth-on-sm: 10ch;
-    --pf-u-min-width--MinWidth-on-md: 20ch;
-  }
-  &__content {
-    --pf-c-page__main-section--PaddingTop: 0;
-  }
-}
